pyproject.toml
[tool.poetry]
name = "troncos"
version = "4.0.0b1"
description = "Collection of Python logging, tracing and profiling tools"
authors = [
    "Guðmundur Björn Birkisson <gudmundur.birkisson@oda.com>",
    "Karl Fredrik Haugland <karlfredrik.haugland@oda.com>",
    "Eirik Martiniussen Sylliaas <eirik.sylliaas@oda.com>",
]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/kolonialno/troncos"
repository = "https://github.com/kolonialno/troncos"
documentation = "https://github.com/kolonialno/troncos"
keywords = ["logs", "traces", "opentelemetry"]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
ddtrace = "1.13.3"
opentelemetry-sdk = "1.17.0"
opentelemetry-exporter-otlp-proto-http = "1.17.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.1.0"
flake8 = "^6.0.0"
mypy = "^1.0.1"
pytest = "^7.2.1"
pytest_codeblocks = "^0.16.1"
fastapi = ">=0.92,<0.96"
requests = "^2.28.2"
pytest-cov = "^4.0.0"
types-requests = "^2.28.11.14"
pytest-asyncio = ">=0.20.3,<0.22.0"
django-environ = "^0.9.0"
pytest-benchmark = "^4.0.0"
snakeviz = "^2.2.0"
django = "^4.2.1"
starlette = "^0.27.0"
structlog = "^23.1.0"


[tool.black]
target-version = ['py311']
line-length = 88
include = '\.pyi?$'
safe = true

[tool.isort]
profile = "black"
skip_gitignore = true
combine_as_imports = true

[tool.mypy]
python_version = "3.11"
files = ["./troncos", "./tests",]
namespace_packages = true
show_error_codes = true
ignore_missing_imports = true
follow_imports = "normal"
strict = true

[[tool.mypy.overrides]]
module = [
    "opentelemetry.sdk.*",
    "opentelemetry.context.*",
]
implicit_reexport = true

[tool.pytest.ini_options]
minversion = 6.0
python_files = "test_*.py"
python_functions = "test_*"
cache_dir = ".pytest_cache"
# Silence warning by opting in to use pytest 6's new default value
junit_family = "xunit2"
# Include captured log messages in system-out in Jenkins report file
junit_logging = "system-out"
console_output_style = "classic"
empty_parameter_set_mark = "xfail"
log_cli_level = "INFO"
log_level = "INFO"
filterwarnings = ["ignore::DeprecationWarning"]

[build-system]
requires = ["poetry-core>=1.1.0a6"]
build-backend = "poetry.core.masonry.api"
